Remove My Folder And Merge Files
