+++ assets/src/js/admin/vendor-edit.js
