+++ src/Admin/views/html-admin-page-vendors-manager.php
