+++ src/VendorHandler.php
